API_django/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
jdatetime==4.1.0
psycopg2-binary
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3
gunicorn==20.1.0
dj_rest_auth
django-allauth
djangorestframework-simplejwt
django-cors-headers==3.13.0
rasterio
Shapely==1.8.2
geojson==2.5.0
